Change package:// url handling to use package: protocol
This change removes the package:// url re-writing in favor of trying to load the package:// url using desktop protocol handlers. This preserves the urls as package:// urls which makes them show up as package:// urls in error messages.

To support .tiff files, this change adds a custom Threejs loader which converts the tiff file to a support Threejs format.

Fixes: #2929